requirements.txt
flask
flask_dropzone
opencv-python
opencv-contrib-python
matplotlib
Pillow